feat: breaking down, moving to buffer from
